requirements.txt
streamlit
pandas
numpy
plotly
yfinance
python-dotenv
requests
openai
alpaca-py
sqlalchemy
psycopg2-binary
